mslearn-ai-document-intelligence/Labfiles/01-prebuild-models/Python/requirements.txt
azure-ai-formrecognizer==3.3.3
python-dotenv>=1.0.0
